LayoutTests/platform/ios-wk1/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/unloading-documents/001-expected.txt

main frame - has 1 onunload handler(s)
frame "<!--frame1-->" - has 1 onunload handler(s)

FAIL document.open in unload assert_equals: expected "0123456789" but got "012389"
